feat(Ueberauth.Strategy.Procore): strategy for Procore login
This strategy is based off of the GitHub strategy, with all of the
customization and return types modified to reflect Procore. Since
Procore requires HTTPS for the callback URL, the `redirect_uri` can be
configured so that a proxy can be used. If not set then the default
generated callback URL will be used.